% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_feeding_behaviour.R
\name{f_behaviour_daily_summary}
\alias{f_behaviour_daily_summary}
\title{Feeding Behaviour Daily Summary}
\usage{
f_behaviour_daily_summary(
  df_in,
  col_animal_id = .data$animal_id,
  col_date = .data$date,
  col_intake = .data$selected_final_intake_kg,
  col_duration = .data$selected_final_duration_sec
)
}
\arguments{
\item{df_in}{A data frame, typically the final output from cleaning pipeline, using `selected_final_duration_sec` and `selected_final_intake_kg`}

\item{col_animal_id, col_date, col_intake, col_duration, col_rate_of_intake}{Column names for the columns in df_in that should be used. Defaults to column names used in Shiny pipeline.}
}
\value{
A dataframe with a row for each date and animal, and columns with new calculated summary values
}
\description{
Data grouped by date and animal id, then summary behaviours calculated.
}
\details{
Calculated behaviour values, for each day:

* intake (kg/d)
* time at feeder (min/d)
* number visits to feeder per d
* mean time per visit (min)
* mean visit size (g)
* mean eating rate (g/min)

Note: The mean eating rate is the mean of each individual event's rate of intake (g/min). This is considered more correct than using the daily mean visit size / mean duration.
}
