% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_bin.R
\name{f_by_bin_clean}
\alias{f_by_bin_clean}
\title{Cleaning - row-by-row within feed bin}
\usage{
f_by_bin_clean(
  df_in,
  zero_thresh = 0.3,
  feedout_thresh = 10,
  col_bin_ID = .data$feed_bin_id,
  col_date = .data$date,
  col_start_time = .data$start_time,
  col_start_weight = .data$start_weight,
  col_end_weight = .data$end_weight,
  col_intake = .data$intake,
  log = TRUE
)
}
\arguments{
\item{df_in}{Data frame to clean.}

\item{zero_thresh}{Number. What threshold (+/- ____ kg) should be used to
determine an intake is about 0 kg? Default +/- 0.3 kg. For example, an
intake would be considered 'negative' if it was <= 0.3 kg, rather than <= 0
kg.}

\item{feedout_thresh}{Number. What threshold (kg) should be used to call an
increase in feed intake a feed out event? Default 10 kg}

\item{col_bin_ID}{Name of column with feed bin ID.}

\item{col_date}{Name of column with date (formatted as Date) of feeding event}

\item{col_start_time}{Name of column with start time (formatted as POSIXct
date-time) recorded by feed bin.}

\item{col_start_weight, col_end_weight}{Name of columns with start and end
weights (in kg) recorded by feed bin.}

\item{col_intake}{Name of column with the feed intake (kg) for the feeding
event, normally calculated as end_weight - start_weight.}

\item{log}{Boolean. Should log files be generated?}
}
\value{
A list of data frames: `df_0kg`, `df_step2_errors` and `df_cleaned`.
In addition, `log_path` is returned in the list, which is the temporary file path where the log is stored.
}
\description{
This script executes a series of individual functions that check for row-by-row
errors in automatically collected feed intake data.
}
\details{
Most steps group the data by day and by feed bin ID, meaning that each day
and feed bin is analysed separately. The final step checks for incorrectly
recorded end times and uses the data grouped by cow ID to evaluate if she
visited another feed bin while still being recorded at another feed bin.
It also generates a log file, summarising each error by count and %. This is
stored in a temporary file path which is returned in the output list.
}
