% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_feeding_behaviour.R
\name{f_combined_meal_analysis}
\alias{f_combined_meal_analysis}
\title{Combined Meal Analysis}
\usage{
f_combined_meal_analysis(
  df_in,
  meal_interval_criteria_min = 5,
  col_animal_id = .data$animal_id,
  col_start_time = .data$start_time,
  col_intake = .data$selected_final_intake_kg,
  col_duration = .data$selected_final_duration_sec
)
}
\arguments{
\item{df_in}{A data frame, typically the final output from cleaning pipeline, using `selected_final_duration_sec` and `selected_final_intake_kg`, and corrected times.}

\item{col_animal_id, col_intake, col_duration}{Column names for the columns in df_in that should be used. Defaults to column names used in Shiny pipeline.}

\item{col_start_time}{Column with start time. NOTE: The end_time is calculated based on the `col_duration` provided so that the corrected duration can be used.}

\item{meal_interval_criteria}{Numeric value defining the time interval (in minutes) to distinguish separate meals based on time lapse between feeding events.}
}
\value{
A dataframe with each row as a 'meal' (aggregate of all feeding events within that meal)
}
\description{
Performs analysis by defining meals based on time intervals between feeding events and then summarizing these meals.
The identification of meals is based on a specified interval criteria, distinguishing separate meals by the time lapse between feeding events.
The `meal_interval_criterion_min` defaults to 5 minutes, meaning that if the interval between a feeding event and the previous event is > 5 minutes it will be defined as a new 'meal'.
Then, all feeding events within the same meal are summarised together in the calculated table.
}
\details{
Calculated meal values include:

* Number of feeding events per meal
* Total meal duration (min)
* Meal size (g)
* Mean eating rate (g/min)
* Meal start and end times
* Total meal duration (sec and min)
* Meal start date
}
