% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_feeding_behaviour.R
\name{f_daily_meal_summaries}
\alias{f_daily_meal_summaries}
\title{Daily Meal Summaries}
\usage{
f_daily_meal_summaries(
  df_in,
  col_animal_id = .data$animal_id,
  col_meal_start_date = .data$meal_start_date,
  col_meal_size_g = .data$meal_size_g,
  col_meal_duration_min = .data$meal_duration_at_bin_min,
  col_meal_total_duration_min = .data$meal_total_duration_min,
  col_meal_mean_eating_rate_g_min = .data$meal_mean_eating_rate_g_min
)
}
\arguments{
\item{df_in}{A data frame containing individual meal data.}

\item{col_animal_id, col_meal_start_date, col_meal_size_g, col_meal_duration_min, col_meal_total_duration_min, col_meal_mean_eating_rate_g_min}{Column names for the
columns in df_in that should be used. Defaults to column names used in Shiny pipeline.}
}
\value{
A dataframe with aggregated daily meal summaries for each animal, detailing the number of meals, and the average and variability of meal sizes, durations, and eating rates.
}
\description{
Summarizes meal data on a daily basis for each animal, aggregating meal data by animal ID and meal start date.
}
\details{
Calculated daily values include:

* Number of meals per day
* Average meal size (g/day)
* Average meal duration at bin (min/day)
* Standard deviation of meal duration at bin
* Average total meal duration (min/day)
* Standard deviation of total meal duration
* Average meal rate of intake (g/min)
* Standard deviation of meal rate of intake

These summaries provide an overview of feeding behavior, highlighting variations in meal size, duration, and intake rates across different days.
}
