% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_cow.R
\name{f_flag_and_replace_outliers}
\alias{f_flag_and_replace_outliers}
\title{f_flag_and_replace_outliers}
\usage{
f_flag_and_replace_outliers(
  df_in,
  col_intake,
  col_duration,
  col_cow_id,
  col_bin_id,
  col_start_time,
  col_date,
  sd_thresh = 5
)
}
\arguments{
\item{df_in}{data frame of one cow to be cleaned.}

\item{col_intake}{name of column with intake data to use, normally the corrected intake from 'by bin' functions}

\item{col_duration}{name of column with duration data to use, normally the corrected duration from 'by bin' functions}

\item{col_cow_id, col_bin_id, col_start_time, col_date}{Column names for cow_id, feed_bin_id, start_time and date from df_in. Used for ID.}

\item{sd_thresh}{number to use as threshold for +/- scale (SD) of fitted bisector regression}
}
\value{
A nested data frame with a column called `data` that is the original data given to equation, and `fitted` which contains
the output from the fitting processes. These can be merged easily into a flat data frame using `f_merge_corrected_outlier_data()`
}
\description{
This function executes a 2 step process for outlier detection: \cr\cr
1) A robust linear model is fit using `MASS::rlm` which produces a rough regression
that is not sensitive to large outliers. The `sd_thresh` set in this function (default of 5 SD)
is used to identify points with a residual >= this threshold and marks them as 'outliers'. \cr
2) Then, a bisector regression is fit to the data that does not include the flagged outliers. This is a
symmetrical linear model and can be used to predict both the x- and y- axis. Therefore, the outliers points
that have a positive residual are re-fitted using this equation to provide a new intake (y value). Likewise,
outliers with a negative residual are re-fitted to provide a new duration (x value).
}
