% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_animal.R
\name{f_merge_corrected_outlier_data}
\alias{f_merge_corrected_outlier_data}
\title{Unnest and merge nested df from `f_flag_and_replace_outliers`}
\usage{
f_merge_corrected_outlier_data(nested_df)
}
\arguments{
\item{nested_df}{A data frame nested by 'fitted' and 'data'}
}
\value{
a single, unnested data frame of fitted and original data
}
\description{
Unnest and merge nested df from `f_flag_and_replace_outliers`
}
