% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_bin_steps.R
\name{f_step2}
\alias{f_step2}
\title{By Bin - Step 2 - Check row-by-row errors}
\usage{
f_step2(
  df_in,
  col_bin_ID,
  col_date,
  col_start_time,
  col_start_weight,
  col_end_weight,
  col_intake
)
}
\arguments{
\item{df_in}{Data frame to clean with the following columns:}

\item{col_bin_ID}{Name of column with feed bin ID.}

\item{col_date}{Name of column with date (formatted as Date) of feeding event}

\item{col_start_time}{Name of column with start time (formatted as POSIXct
date-time) recorded by feed bin.}

\item{col_start_weight, col_end_weight}{Name of columns with start and end
weights (in kg) recorded by feed bin.}

\item{col_intake}{Name of column with the feed intake (kg) for the feeding
event, normally calculated as end_weight - start_weight.}
}
\value{
A list of 3 data frames is returned. 1) `step2_full` - with no rows
  filtered, 2) The errors are removed and stored in `step2_errors`, and 3)
  all rows that are `ok` are returned in `step2_ok` data frame
}
\description{
This function groups data by date and feed bin ID and sorts by start time
before adding columns with the previous end weight, previous start weight,
next end weight and next start weight. This makes it easy to compare the rows
before and after the current row. Then, it checks the difference between the
previous end weight and next start weight. If this is 0 kg, then it assumes
the current row does not align with the weight data and is therefore an
error. The rows flagged with this error are then classified into either:
"error type:  intake event <= |0.2|" or "error type:  intake event > |0.2|".
This suggests that if the intake is greater than 0.2 or less than -0.2, it
might be useful to manually check this classification.
}
\details{
This was designed to remove rows where the entry appears to be out of place
and typically these entries had a start and end weight that were completely
out of line with the surrounding data, as if the entry had been misplaced
from another time or feed bin. Without knowing how to deal with this, or if
it is even a real event, it is easiest to remove. Doing so also improves the
ability of this script to detect row-by-row errors.
}
