% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_bin_steps.R
\name{f_step6_correct_end_times}
\alias{f_step6_correct_end_times}
\title{By Bin - Step 6 - Correct end times}
\usage{
f_step6_correct_end_times(
  df_step5,
  col_animal_id = .data$animal_id,
  col_start_time = .data$start_time,
  col_end_time = .data$end_time
)
}
\arguments{
\item{df_step5}{Data frame produced by [f_step5()]}

\item{col_animal_id}{Name of column with animal ID.}

\item{col_start_time, col_end_time}{Name of column with start and end time (formatted as POSIXct
date-time) recorded by feed bin.}
}
\value{
The original data frame with the following additional columns:
nextStartTime, is_end_time_overlap_error, corrected_end_time, corrected_duration_sec_diff, corrected_duration_sec
}
\description{
This function checks for errors in the time recorded by the feed bin without making
any assumptions about a biologically possible rate of intake. This is technically
a 'by animal' step, as it compares start and end times within a animal (across feed bins)
to see if the animal visited another feed bin while still being recorded as present
at a feed bin (i.e. overlapping events). If this is found, it will trim the first
event to end 10 sec before the next recorded feeding event.
}
