% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploads_utils.R
\name{fct_animal_icon}
\alias{fct_animal_icon}
\title{Draw SVG animal face}
\usage{
fct_animal_icon(col)
}
\arguments{
\item{col}{colour to use for lines. either 'black' or 'white'}
}
\value{
returns HTML() with the <svg> code to draw icon.
Can be tested in R console using `htmltools::browsable(fct_animal_icon('white'))`
}
\description{
Draw SVG animal face
}
