% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_animal_utils.R
\name{fct_check_for_valuebox_animal}
\alias{fct_check_for_valuebox_animal}
\title{Check errors for value box}
\usage{
fct_check_for_valuebox_animal(df_summary, col, condition)
}
\arguments{
\item{df_summary}{summary table that has a column with different errors}

\item{col}{What is the name of the column that was counted}

\item{condition}{A string to match which error's count data should be returned}
}
\value{
A number to print in value box
}
\description{
Helper function to return the value of that count for when it meets a certain condition.
Assumes has been counted with [f_count_errors()].
}
