% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_bin_utils.R
\name{fct_check_for_valuebox_bin}
\alias{fct_check_for_valuebox_bin}
\title{Count and check errors for valuebox}
\usage{
fct_check_for_valuebox_bin(df, col, condition)
}
\arguments{
\item{df}{a data frame with at least col}

\item{col}{column name to check}

\item{condition}{A string to match which error's count data should be returned}
}
\value{
a number to print in value box
}
\description{
Helper function to count a data frame by a particular column, and then
return the value of that count for when it meets a certain condition.
}
