% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploads_utils.R
\name{fct_date_time_parse_error_handler}
\alias{fct_date_time_parse_error_handler}
\title{Error handling for date / times}
\usage{
fct_date_time_parse_error_handler(expr)
}
\arguments{
\item{expr}{A section of R code inside of a `{}`, see [tryCatch()]}
}
\value{
Returns NULL because it can't keep the data frame with incorrect files
}
\description{
This is a wrapper for tryCatch that is designed to be used for checking the format
of dates. Normally, lubridate will throw a warning if it can't format the dates/times correctly.
This function will detect that error and create a modal to warn the user that the df was uploaded, but that the
dates/times are likely incorrect.
}
\details{
There is a slight performance decrease as the expr is evaluated a second time to return the data.
}
