% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_all_mods.R
\name{fct_DT_nopages}
\alias{fct_DT_nopages}
\title{DT table template}
\usage{
fct_DT_nopages(
  x,
  scrollY = 350,
  buttons_at_bottom = FALSE,
  rownames = FALSE,
  fillContainer = FALSE
)
}
\arguments{
\item{x}{dataframe}

\item{scrollY}{number to represent size of y axis before scrolling}

\item{buttons_at_bottom}{boolean. Should buttons for downloading table be at bottom?}

\item{rownames}{boolean. Are rownames used in this table? Default FALSE.}

\item{fillContainer}{boolean for DT::datatable(fillContainer)}
}
\value{
DT
}
\description{
Call to `DT::datatable()` with added download buttons, fixed columns and no pages
}
