% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_all_mods.R
\name{fct_DT_pages}
\alias{fct_DT_pages}
\title{DT table template - for using pages}
\usage{
fct_DT_pages(
  x,
  pageLength = 20,
  scrollY = 350,
  allow_copy = FALSE,
  buttons_at_bottom = FALSE,
  rownames = FALSE,
  fillContainer = FALSE
)
}
\arguments{
\item{x}{dataframe}

\item{pageLength}{number of rows to show per page.}

\item{scrollY}{number to represent size of y axis before scrolling}

\item{allow_copy}{bool. Should the option to copy data be enabled? For large data it is not advisable.}

\item{buttons_at_bottom}{boolean. Should buttons for downloading table be at bottom?}

\item{rownames}{boolean. Are rownames used in this table? Default FALSE.}

\item{fillContainer}{boolean for DT::datatable(fillContainer)}
}
\value{
DT
}
\description{
Call to `DT::datatable()` with added download buttons, fixed columns and uses pages. Pages are helpful when large data needs to be loaded. It also allows scrolling.
}
