% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploads_utils.R
\name{fct_import_DAT_default}
\alias{fct_import_DAT_default}
\title{Import DAT files with default column names}
\usage{
fct_import_DAT_default(.x, .y)
}
\arguments{
\item{.x}{datapath from import$DAT_in$datapath}

\item{.y}{filename from import$DAT_in$name}
}
\description{
These are hard-coded column names, expecting a specific file type. It also
requires the file names to have the date as 6 numbers in it as: YYMMDD.
This is the default behaviour of DAT files from insentec.
}
