% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploads_utils.R
\name{fct_missing_col_error_handler}
\alias{fct_missing_col_error_handler}
\title{Error handling for missing columns}
\usage{
fct_missing_col_error_handler(expr)
}
\arguments{
\item{expr}{A section of R code inside of a `{}`, see [tryCatch()]}
}
\value{
Returns the result of expr, which is normally a data frame.
}
\description{
This is a wrapper for tryCatch that is designed to be used for checking if columns are missing.
Normally, select(all_of()) will throw a meainingful warning.
This function will detect that error and create a modal to warn the user that the upload failed.
}
