% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_bin_utils.R
\name{fct_modal_content_bybin_flowchart}
\alias{fct_modal_content_bybin_flowchart}
\title{Function to create HTML content for modal in mod_by_bin_clean.R}
\usage{
fct_modal_content_bybin_flowchart(svgContent)
}
\arguments{
\item{svgContent}{SVG content object. Normally loaded into session with a
call to readLines(con = app_sys("app/www/filename.svg"), warn = FALSE)}
}
\value{
a list of HTML
}
\description{
Function to create HTML content for modal in mod_by_bin_clean.R
}
