% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_bin_utils.R
\name{fct_modal_content_bybin_more_info}
\alias{fct_modal_content_bybin_more_info}
\title{Function to store text for 'more info' button for mod_by_bin
Normally parsed to fct_show_custom_modal()}
\usage{
fct_modal_content_bybin_more_info()
}
\value{
a list of HTML
}
\description{
Function to store text for 'more info' button for mod_by_bin
Normally parsed to fct_show_custom_modal()
}
