% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_cow_utils.R
\name{fct_plot_by_animal_overall}
\alias{fct_plot_by_animal_overall}
\title{Plot for overall data}
\usage{
fct_plot_by_animal_overall(df, col_intake, col_duration, pt_size = 3)
}
\arguments{
\item{df}{data frame to plot}

\item{col_intake}{column with intake data (y axis)}

\item{col_duration}{column with duration data (x axis)}

\item{pt_size}{point size to use (for ggplot2)}
}
\value{
a ggplot2
}
\description{
By animal overall plot which plots a regression showing the different outliers.
Assumes various columns are present from the [f_flag_and_replace_outliers()] function
}
