% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_cow_utils.R
\name{fct_plot_by_cow}
\alias{fct_plot_by_cow}
\title{fct_plot_by_cow}
\usage{
fct_plot_by_cow(
  df,
  max_intake_rate_kg_min = 1.5,
  min_intake_rate_kg_min = 0.05,
  outlier_exemption_max_duration = 1,
  outlier_exemption_max_intake = 0.3,
  col_intake = .data$corrected_intake_bybin,
  col_duration = .data$corrected_duration_sec_seconds,
  col_animal_id = .data$animal_id,
  pt_size = 3
)
}
\arguments{
\item{df}{dataframe to plot}

\item{min_intake_rate_kg_min, max_intake_rate_kg_min}{number (kg/min). Events with a rate of intake (i.e. intake / duration) greater than the `max_intake_rate_kg_min` and less than the `min_intake_rate_kg_min` will be classed as a 'manual outlier'}

\item{outlier_exemption_max_duration, outlier_exemption_max_intake}{number (kg or min). Events below this `outlier_exemption_max_intake` (kg) & with a duration less than `outlier_exemption_max_duration` (min) are exempt from other outlier detection methods.}

\item{col_intake}{name of column with intake data to use}

\item{col_duration}{name of column with duration data to use}

\item{col_animal_id}{name of column with cow ID}

\item{pt_size}{number. point size for `geom_point()`}
}
\value{
ggplot2 object
}
\description{
Plot 'by cow' outlier detection results. Visualise feed duration vs intake
for individual cows. Each point is an individual feeding event. Outlier
points are shown twice, with an arrow pointing from original value to new
fitted value. The outlier classifications are based on values selected when
cleaned. Outlier type refers to negative (neg) or positive (pos) residual
from fitted regression. Boundary lines are the max and min rates of intake
selected when cleaning to define outliers. The red box represents the
user-defined region where values are 'exempt' from outlier detection. Select
'interactive' plot on left to allow more data on hover or to zoom in and out.
}
\details{
Assumes certain column names that were generated by `f_flag_and_replace_outliers`: 'new_x', 'new_y', 'predicted_y_bisector', 'is_outlier', 'outlier_pos_neg', 'animal_id'
}
