% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_all_mods.R
\name{fct_plots_summary}
\alias{fct_plots_summary}
\title{Summary plots}
\usage{
fct_plots_summary(
  df_in,
  x,
  y = NULL,
  col_colour = .data$check_end_weights,
  type
)
}
\arguments{
\item{df_in}{data frame to plot}

\item{x}{column name for x axis (if regression) or for histogram}

\item{y}{if regression, then name of column for y axis (typically intake)}

\item{col_colour}{name of column to colour regression by}

\item{type}{either 'hist' or 'reg' for histogram or regression, respectively}
}
\value{
a ggplot2
}
\description{
Summary plots of either regression or histogram
}
