% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_animal_utils.R
\name{fct_Rd_to_HTML}
\alias{fct_Rd_to_HTML}
\title{Convert Rd to HTML}
\usage{
fct_Rd_to_HTML(Rd_file)
}
\arguments{
\item{Rd_file}{An .Rd file, can be from app_sys(paste0("man/", filename)) or parsed directly e.g. from tools::Rd_db('IntakeInspectR')}
}
\value{
HTML file
}
\description{
This is used to convert the documentation .Rd files into HTML that can be
used inside a modal.
}
\details{
For using in a modal, this function will be called inside a renderUI() that
is assigned to an output$ . Then, this output is called inside a
`showModal(modalDialog( uiOutput(HERE)))`
}
