% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intake4Rumi-package.R
\docType{package}
\name{IntakeInspectR-package}
\alias{IntakeInspectR}
\alias{IntakeInspectR-package}
\title{IntakeInspectR: Data cleaning and visualisation for automatically collected feed intake data}
\description{
A set of tools to help researchers clean their raw feed intake data by checking for common errors and visualise their data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CNM-University-of-Guelph/IntakeInspectR}
  \item Report bugs at \url{https://github.com/CNM-University-of-Guelph/IntakeInspectR/issues}
}

}
\author{
\strong{Maintainer}: Dave Innes \email{innesd@uoguelph.ca}

}
\keyword{internal}
