% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_bin_plots.R
\name{plot_bin_timeline}
\alias{plot_bin_timeline}
\title{Plot timeline of feeding events at each feed bin}
\usage{
plot_bin_timeline(
  df_in,
  col_start_time,
  col_end_time,
  col_start_weight,
  col_end_weight,
  col_colour,
  col_hover_text
)
}
\arguments{
\item{df_in}{Data frame that has been filtered via user input (`input$bin_id`) to only include 1 feed bin.}

\item{col_start_time, col_end_time, col_start_weight, col_end_weight, col_colour}{Column names used for plotting}

\item{col_hover_text}{column name of extra text for tooltip. Usually from adding a column with paste, e.g. `paste('cow_id:', cow_id)`}
}
\value{
A ggplot2 object
}
\description{
A function to plot a point for start and end of feeding event, connected by a line, for each feed bin.
}
