% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_by_bin.R
\name{f_by_bin_bypass_shiny}
\alias{f_by_bin_bypass_shiny}
\title{Bypass 'By Bin' cleaning for Shiny app}
\usage{
f_by_bin_bypass_shiny(df_in, log = TRUE)
}
\arguments{
\item{df_in}{Data frame that was uploaded by user}

\item{log}{bool. Should log be created.}
}
\value{
A list of data frames: `df_0kg`, `df_step2_errors` and `df_cleaned`.
In addition, `log_path` is returned in the list, which is the temporary file path where the log is stored.
}
\description{
This will return a similar output to f_by_bin_clean that will allow a user to bypass this step in the Shiny app.
}
