% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_feeding_behaviour.R
\name{f_final_daily_summary}
\alias{f_final_daily_summary}
\title{Final Daily Summary}
\usage{
f_final_daily_summary(df_daily_summary, df_daily_meal_summaries)
}
\arguments{
\item{df_daily_summary}{A dataframe containing daily behavioral summary data for each animal. Normally calculated by [f_behaviour_daily_summary()]}

\item{df_daily_meal_summaries}{A dataframe containing daily meal summary data for each animal. Normally calculated by [f_daily_meal_summaries()]}
}
\value{
A dataframe that combines the daily behavioral summaries with daily meal summaries for each animal, matched by animal ID and date.
}
\description{
Merges daily behavioral summaries with daily meal summaries into a single
comprehensive dataset.
}
