% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_cow_utils.R
\name{fct_plot_by_cow}
\alias{fct_plot_by_cow}
\title{fct_plot_by_cow}
\usage{
fct_plot_by_cow(df, col_intake, col_duration, pt_size = 3)
}
\arguments{
\item{df}{dataframe to plot}

\item{col_intake}{name of column with intake data to use}

\item{col_duration}{name of coumn with duration data to use}

\item{pt_size}{number. point size for `geom_point()`}
}
\value{
ggplot2 object
}
\description{
Plot outlier detection results
}
\details{
Assumes certain column names that were generated by `f_flag_and_replace_outliers`: 'new_x', 'new_y', 'predicted_y_bisector', 'is_outlier', 'outlier_pos_neg', 'cow_id'
}
